Darwin: Implement MSYS path translation hack
Changed configure of gcc/llvmgcc to:
--with-gxx-include-dir=./usr/include/c++/4.2.1 i.e. adding a '.' to
prevent MSYS path translation from mangling this path. This is then
accounted for and fixed up in fix_standard_paths in c-incpath.c.

It is likely that MSYS2 will provide a specific prefix (such as
donttranslate://) to prevent this mangling from happening, and once
this is in place, the *-msys-path-translation-hack.patch files will
be removed.