GCC: Adopt POSIX semantics for include dir existence
On Windows (MinGW-w64 *and* Cygwin) non-existant dirs
that are stepped out of via '..' are not considered a
problem and files can be returned to live in them via
open(). If any '..' are found in the path then take a
a more thorough code-path which checks each directory
in the chain exists.

Patch supplied for GCC versions 4.8.5, 4.9.3 and
5.3.0

Signed-off-by: Ray Donnelly <mingw.android@gmail.com