arch/all: Add common function to return multilib target
This code was abstracted out of Cody P Schafer's multilib patch.
It doesn't seem right having architecture dependent code in something generic
like the (e)glibc-common script. So this patch breaks it out into
scripts/build/arch/<arch>.sh in a function:

  multilib_target_to_build="$(CT_DoArchMultilibTarget 'multi_flags' 'target-in')"

Alexey Neyman had an that we should query GCC directly for the multiarch flags,
so it's worth investigating this more. I think it would involve --enable-multiarch
From gcc.gnu.org/install/configure.html

"
--enable-multiarch
Specify whether to enable or disable multiarch support. The default is to check
for glibc start files in a multiarch location, and enable it if the files are
found. The auto detection is enabled for native builds, and for cross builds
configured with --with-sysroot, and without --with-native-system-header-dir.
"

Signed-off-by: Bryan Hundven <bryanhundven@gmail.com>
Signed-off-by: Ray Donnelly <ray.donnelly@gmail.com>