configure: prefer g-variants
For awk and make, we currently check the standard names as found on
a classic GNU-userland system, and then fallback to looking for the
g-named variants as found on a *BSD system.

But it happens that the usual name might also be available on a *BSD
system, so we should first check the g-variant.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>