functions: Allow patch directories to be files
On Windows (Cygwin or MSYS2) without native symlinks enabled,
Git writes out symlink directories as files containing the target
as text.

So if a patch directory is actually a file, then dereference it
(only once) and then continue the normal logic of checking that
it is a directory.

glibc/linaro-2.20-* is a symlink to glibc/2.20, and this is enough
to make this work with Windows as a build system.

Symlinks require special permissions on Windows unfortunately ..

Signed-off-by: Ray Donnelly <mingw.android@gmail.com>