Add POST_SYSROOT_PREFIX
usr isn't always wanted, in fact on Windows
it can break us.

At the same time, fix up the mingw layout to be
the same as Linux. If you want the old layout,
POST_SYSROOT_PREFIX can be used to achieve that
(set it to e.g. "mingw/x86_64-w64-mingw32")

.. there is more work to do here for mingw-w64:
"mingw" is hardcoded in the target header file and
it should be configureable. Sane examples would be
"usr", "mingw32" and "mingw64".

This patch is not final. I think the whole
SYSROOT situation in crosstool-ng is a mess
and I would really like to untangle it.