linux/glibc: Add CentOS5 versions (2.6.18.4/2.5.1+2.5??)
It is unclear whether I should be using 2.5 or 2.5.1 here as I
cannot remember why I have both or if both sets apply cleanly.

I *think* 2.5 should be getting used and 2.5.1 should be fixed up
or removed, but either way not used for CentOS5.11.

Adds the huge combo patch (kernel-2.6.18-redhat.patch) from RedHat
which is in this archive:

http://vault.centos.org/5.11/updates/SRPMS/kernel-2.6.18-419.el5.src.rpm